AIGRID, ELAS, HFA: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31633 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/aigrid/aigccitt.c
@@ -1539,7 +1539,7 @@ typedef struct {
     sp->bit = BitsAvail;						\
     sp->data = BitAcc;							\
     sp->EOLcnt = EOLcnt;						\
-    rawcc -= (unsigned char *) cp - rawcp;				\
+    rawcc -= (int)((unsigned char *) cp - rawcp);			\
     rawcp = (unsigned char *) cp;					\
 } while (0)
